Beverly {startrek}	R Documentation

Talk like Commander Beverly Cheryl Crusher, MD, (née Howard)

Description

Talk with the sensible and kind insight that a Chief Medical Officer on a starship develops.

Usage

Beverly()

Value

a phrase

See Also

Other Darmok: Computer, Data, Dathon, Geordi, Guinan, Picard, Pulaski, Q, Riker, Troi, Wesley, Worf

[Package startrek version 0.1.0 Index]