Geordi {startrek}	R Documentation

Speak like Geordi

Description

Truly man with a vision. Geordi can see in many frequencies. Expect a lot of technobabble and a few creepy romance words.

Usage

Geordi()

Value

a phrase

See Also

Other Darmok: Beverly, Computer, Data, Dathon, Guinan, Picard, Pulaski, Q, Riker, Troi, Wesley, Worf

Examples

Geordi()

[Package startrek version 0.1.0 Index]