Picard {startrek}	R Documentation

Return a phrase that Picard uttered

Description

Be as eloquent as our favorite archaeologist, diplomat and decorated Starfleet officer Jean-Luc Picard, or as we tend to say "JLP". The startrek package does not help you in becoming a Starfleet officer, or archaeologist for that matter. But perhaps you will use the same phrasing when you order your Tea Earl Grey. As Picard once said : "Darmok and Jalad... at Tanagra".

Usage

Picard()

Value

a phrase

See Also

Other Darmok: Beverly, Computer, Data, Dathon, Geordi, Guinan, Pulaski, Q, Riker, Troi, Wesley, Worf

Examples

Picard()

[Package startrek version 0.1.0 Index]