Riker {startrek}	R Documentation

William (Will) "Number One" Riker

Description

The original leaning man. Riker has a tendency to solve diplomatic disputes by sleeping with anyone around. The man, the legend.

Usage

Riker()

Value

a phrase

See Also

Other Darmok: Beverly, Computer, Data, Dathon, Geordi, Guinan, Picard, Pulaski, Q, Troi, Wesley, Worf

Examples

Riker()

[Package startrek version 0.1.0 Index]