random_tng_character {startrek}	R Documentation

Call a random phrase from Q, Worf, Wesley, Picard, Data or Guinan

Description

Usefull if you want a phrase at the end of a long process.

Usage

random_tng_character()

Examples

random_tng_character()

[Package startrek version 0.1.0 Index]