startrek {startrek}	R Documentation

Star trek package write like characters spoke

Description

This incredibly unuseful package makes you talk like the main characters from Star Trek The Next Generation. call functions with the name of the characters 'Picard()' , 'Geordi()', etc Characters included are Picard, Beverly, Data, Geordi, Guinan, Computer, Pulaski, Q, Riker, Troi, Wesley, Worf and also Dathon.

[Package startrek version 0.1.0 Index]